refactor(js/timestamp): flatten showcase files, outsource agent executor, and rename fixed time variable
