+++ samples/js/extensions/timestamp/package.json
@@ -4,8 +4,8 @@
   "description": "Timestamp Extension sample using @a2a-js/sdk v1.0",
   "type": "module",
   "scripts": {
-    "start": "tsx tests/index.ts",
-    "client": "tsx tests/test_client.ts"
+    "start": "tsx index.ts",
+    "client": "tsx test_client.ts"
   },
   "dependencies": {
     "@a2a-js/sdk": "^1.0.0-beta.0",
